requirements.txt

disnake==2.9.0
Pillow==9.5.0
python-dotenv==1.0.0
python_Levenshtein==0.21.1
